+++ packages/ossa/src/components/radio/index.tsx
@@ -1,7 +1,8 @@
 import React from "react";
 import { View } from "@tarojs/components";
 import classNames from "classnames";
-import { OsRadioProps } from "../../../types/index";
+import RadioOption from "./option/index";
+import { OsRadioProps, OsRadioOptionProps } from "../../../types/index";
 
 function getStyleObj(props: OsRadioProps) {
   const _styleObj = {};
@@ -15,10 +16,23 @@ function getClassObject(props: OsRadioProps) {
   return classObject;
 }
 
+function getCommonChildProps(props: OsRadioProps) {
+  const childProps: Pick<OsRadioProps, "value" | "isDisabled" | "isReadonly"> =
+    {};
+
+  ["value", "isDisabled", "isReadonly"].forEach((key) => {
+    if (props[key] !== undefined) {
+      childProps[key] = props[key];
+    }
+  });
+  return childProps;
+}
+
 export default function Index(props: OsRadioProps) {
   const rootClassName = "ossa-radio"; //组件
   const classObject = getClassObject(props); //组件修饰
   const styleObject = getStyleObj(props);
+  const coommonChildProps = getCommonChildProps(props);
 
   return (
     <View
@@ -36,7 +50,19 @@ export default function Index(props: OsRadioProps) {
           props.type === "row" ? "ossa-radio__options--row" : ""
         )}
       >
-        {props.children}
+        {React.Children.map(props.children, (child) => {
+          return React.isValidElement(child) && child.type === RadioOption
+            ? React.createElement(child.type, {
+                ...(child.props as OsRadioOptionProps),
+                ...coommonChildProps,
+                onClick: (v) => {
+                  const onClick = (child.props as OsRadioOptionProps).onClick;
+                  onClick && onClick(v);
+                  props.onChange && props.onChange(v);
+                },
+              })
+            : child;
+        })}
       </View>
     </View>
   );